add Vite as build tool
