+++ index.html
@@ -5,73 +5,51 @@
     <title>Tic Tac Toe</title>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="icon" type="image/x-icon" href="static/icon/tic-tac-toe-game.svg">
-    <link rel="stylesheet" href="static/css/main.css">
+    <link rel="icon" type="image/x-icon" href="public/tic-tac-toe-game.svg">
 </head>
 
 <body>
     <div class="container">
         <h1>TicTacToe</h1>
-        <img src="static/icon/tic-tac-toe-game.svg" class="center" width="64" alt="decorative">
+        <img src="/tic-tac-toe-game.svg" class="center" width="64" alt="decorative">
+        <label for="player1" class="hidden">Player 1</label>
         <input name="player1" id="player1" placeholder="Spieler 1" type="text" required>
         <br>
+        <label for="player2" class="hidden">Player 2</label>
         <input name="player2" id="player2" placeholder="Spieler 2" type="text" required>
         <br>
-        <input type="submit" value="New Game" onclick="setPlayerNames()">
+        <input id="submit" type="submit" value="New Game">
     </div>
     <ul class="background">
         <li>
-            <img src="/static/icon/burger.svg" alt="decorative">
+            <img src="/burger.svg" alt="decorative">
         </li>
         <li>
-            <img src="/static/icon/donut.svg" alt="decorative">
+            <img src="/donut.svg" alt="decorative">
         </li>
         <li>
-            <img src="/static/icon/burger.svg" alt="decorative">
+            <img src="/burger.svg" alt="decorative">
         </li>
         <li>
-            <img src="/static/icon/donut.svg" alt="decorative">
+            <img src="/donut.svg" alt="decorative">
         </li>
         <li>
-            <img src="/static/icon/burger.svg" alt="decorative">
+            <img src="/burger.svg" alt="decorative">
         </li>
         <li>
-            <img src="/static/icon/donut.svg" alt="decorative">
+            <img src="/donut.svg" alt="decorative">
         </li>
         <li>
-            <img src="/static/icon/burger.svg" alt="decorative">
+            <img src="/burger.svg" alt="decorative">
         </li>
         <li>
-            <img src="/static/icon/donut.svg" alt="decorative">
+            <img src="/donut.svg" alt="decorative">
         </li>
         <li>
-            <img src="/static/icon/burger.svg" alt="decorative">
-        </li>
-        <li>
+            <img src="/burger.svg" alt="decorative">
         </li>
     </ul>
 </body>
-<script>
-    console.log(window.location.origin);
-    function setPlayerNames() {
-        let player1 = document.getElementById('player1').value;
-        let player2 = document.getElementById('player2').value;
-        sessionStorage.setItem("player1", player1);
-        sessionStorage.setItem("player2", player2);
-        if (player1 !== "" && player2 !== "") {
-            let inputFields = document.querySelectorAll("input[type=text]");
-            inputFields.forEach(element => {
-                element.style.border = "none";
-            })
-            window.location.href = window.location.origin + "/tictactoe.html";
-        }
-        else {
-            let inputFields = document.querySelectorAll("input[type=text]");
-            inputFields.forEach(element => {
-                element.style.border = "2px solid red";
-            })
-        }
-    }
-</script>
+<script type="module" src="/src/index.js"></script>
 
 </html>
