+++ src/main.css
@@ -16,6 +16,10 @@ body {
   padding: 0;
 }
 
+.hidden {
+  display: none;
+}
+
 @keyframes animate {
   0% {
     transform: translateY(0) rotate(0deg);
