+++ tictactoe.html
@@ -5,10 +5,7 @@
     <title>Tic Tac Toe</title>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="icon" type="image/x-icon" href="static/icon/tic-tac-toe-game.svg">
-    <link rel="stylesheet" href="static/css/main.css">
-    <link rel="stylesheet" href="static/css/tictactoe.css">
-
+    <link rel="icon" type="image/x-icon" href="public/tic-tac-toe-game.svg">
 <body>
     <div class="winner" id="win">Hat gewonnen</div>
     <div class="game">
@@ -16,42 +13,42 @@
         <table class="playground">
             <tr>
                 <td>
-                    <div class="playgroundButton" style="--direction: 1;" x="0" y="0"></div>
+                    <div class="playgroundButton" style="--direction: 1;" data-x="0" data-y="0"></div>
                 </td>
                 <td>
-                    <div class="playgroundButton"  style="--direction: 1;" x="1" y="0"></div>
+                    <div class="playgroundButton"  style="--direction: 1;" data-x="1" data-y="0"></div>
                 </td>
                 <td>
-                    <div class="playgroundButton" style="--direction: -1;" x="2" y="0"></div>
-                </td>
-            </tr>
-            <tr>
-                <td>
-                    <div class="playgroundButton" style="--direction: 1;" x="0" y="1"></div>
-                </td>
-                <td>
-                    <div class="playgroundButton" style="--direction: -1;" x="1" y="1"></div>
-                </td>
-                <td>
-                    <div class="playgroundButton" style="--direction: 1;" x="2" y="1"></div>
+                    <div class="playgroundButton" style="--direction: -1;" data-x="2" data-y="0"></div>
                 </td>
             </tr>
             <tr>
                 <td>
-                    <div class="playgroundButton" style="--direction: -1;" x="0" y="2"></div>
+                    <div class="playgroundButton" style="--direction: 1;" data-x="0" data-y="1"></div>
                 </td>
                 <td>
-                    <div class="playgroundButton" style="--direction: 1;" x="1" y="2"></div>
+                    <div class="playgroundButton" style="--direction: -1;" data-x="1" data-y="1"></div>
                 </td>
                 <td>
-                    <div class="playgroundButton" style="--direction: -1;" x="2" y="2"></div>
+                    <div class="playgroundButton" style="--direction: 1;" data-x="2" data-y="1"></div>
+                </td>
+            </tr>
+            <tr>
+                <td>
+                    <div class="playgroundButton" style="--direction: -1;" data-x="0" data-y="2"></div>
+                </td>
+                <td>
+                    <div class="playgroundButton" style="--direction: 1;" data-x="1" data-y="2"></div>
+                </td>
+                <td>
+                    <div class="playgroundButton" style="--direction: -1;" data-x="2" data-y="2"></div>
                 </td>
             </tr>
         </table>
-        <button class="reset" onclick="resetGame()">Reset game</button>
-        <button class="newGame" onclick="createNewGame()">Create new Game</button>
+        <button class="reset">Reset game</button>
+        <button class="newGame">Create new Game</button>
     </div>
-    <script src="static/js/tictactoe.js"></script>
+    <script type="module" src="src/tictactoe.js"></script>
 </body>
 
 </html>
